refactor: split client into 2 classes
